+++ spec/code/parser/dictionnary_spec.rb
@@ -4,42 +4,15 @@ RSpec.describe Code::Parser::Dictionnary do
   subject { described_class.new.parse(input) }
 
   [
-    [
-      '{name: "Dorian"}',
-      {
-        dictionnary: [{ key: { name: "name" }, value: [{ string: "Dorian" }] }],
-      },
-    ],
-    [
-      '{a: true, "b": false}',
-      {
-        dictionnary: [
-          { key: { name: "a" }, value: [{ boolean: "true" }] },
-          { key: { string: "b" }, value: [{ boolean: "false" }] },
-        ],
-      },
-    ],
-    [
-      "{ true => 1, false => 2}",
-      {
-        dictionnary: [
-          {
-            key: [{ boolean: "true" }],
-            value: [{ number: { base_10: { integer: { whole: "1" } } } }],
-          },
-          {
-            key: [{ boolean: "false" }],
-            value: [{ number: { base_10: { integer: { whole: "2" } } } }],
-          },
-        ],
-      },
-    ],
-  ].each do |(input, expected)|
+    '{name: "Dorian"}',
+    '{a: true, "b": false}',
+    "{ true => 1, false => 2}",
+  ].each do |input|
     context input.inspect do
       let(:input) { input }
 
       it "succeeds" do
-        expect(subject).to eq(expected)
+        expect { subject }.to_not raise_error
       end
     end
   end